+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Master (Unreleased)
 
+# 0.19.1 - 23 July, 2019
+
+### New features:
+
+- Allow to extend knex query builder #3334
+- Add .isCompleted() to transaction #3368
+- Minor enhancements around aliasing of aggregates #3354
+
+### Typings
+
+- Update configuration typings to allow for oracle db connectionstring #3361
+- Update Knex.raw type to be any by default because the actual type is dialect specific #3349
+
 # 0.19.0 - 11 July, 2019
 
 ### Changes: